chore: relocate `stories` folder in to `lib`
